// src/types.ts
export interface IMovieData {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: string;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}

export interface ResultType {
    page: number;
    results: IMovieData[];
    total_pages: number;
    total_results: number;
}

export interface MovieDetailState {
    details: IMovieDetail,
    cast: ICastRoot
}

// ** Movie Detail Interfaces
export interface IMovieDetail {
    adult: boolean
    backdrop_path: string
    belongs_to_collection: any
    budget: number
    genres: IGenre[]
    homepage: string
    id: number
    imdb_id: string
    original_language: string
    original_title: string
    overview: string
    popularity: number
    poster_path: string
    production_companies: IProductionCompany[]
    production_countries: IProductionCountry[]
    release_date: string
    revenue: number
    runtime: number
    spoken_languages: ISpokenLanguage[]
    status: string
    tagline: string
    title: string
    video: boolean
    vote_average: number
    vote_count: number
}

export interface IGenre {
    id: number
    name: string
}

export interface IProductionCompany {
    id: number
    logo_path: string
    name: string
    origin_country: string
}

export interface IProductionCountry {
    iso_3166_1: string
    name: string
}

export interface ISpokenLanguage {
    english_name: string
    iso_639_1: string
    name: string
}


// ** Movie Cast Interfaces

export interface ICastRoot {
    id: number
    cast: ICast[]
    crew: ICrew[]
}

export interface ICast {
    adult: boolean
    gender: number
    id: number
    known_for_department: string
    name: string
    original_name: string
    popularity: number
    profile_path?: string
    cast_id: number
    character: string
    credit_id: string
    order: number
}

export interface ICrew {
    adult: boolean
    gender: number
    id: number
    known_for_department: string
    name: string
    original_name: string
    popularity: number
    profile_path?: string
    credit_id: string
    department: string
    job: string
}